refactor/reorder-components pulled lists dropdown into separate component. fixed bug keeping moviecards from rendering.
